Dev auth

+++ src/views/Article.vue
@@ -1,70 +1,77 @@
 <template>
   <div class="container">
-    <div class="block_topic center media_mobile">
-      <span class="text_topic font_ny_el font_bold">{{ currentPoster.title }}</span>
+    <div v-if="getLoading">
+      <Loader />
     </div>
-    <Poster v-if="visible" v-bind:visible="visible" v-on:closeModal="closeModal" />
-    <div class="block_question center media_mobile">
-      <span class="text_question font_ny_el font_normal"
-        >{{ currentPoster.subtitle }}
-      </span>
-    </div>
+    <div v-else>
+      <div class="block_topic center media_mobile">
+        <span class="text_topic font_ny_el font_bold">{{ currentPoster.title }}</span>
+      </div>
+      <Poster v-if="visible" v-bind:visible="visible" v-on:closeModal="closeModal" />
+      <div class="block_question center media_mobile">
+        <span class="text_question font_ny_el font_normal"
+          >{{ currentPoster.subtitle }}
+        </span>
+      </div>
 
-    <div class="article_image center">
-      <img class="image_banner media_mobile" :src="currentPoster.src" />
-      <figcaption class="caption_text font_sf_text">
-        Image caption centered this way and I’ll make this a bit longer to indicate the
-        amount of line-height.
-      </figcaption>
-    </div>
+      <div class="article_image center">
+        <img class="image_banner media_mobile" :src="currentPoster.src" />
+        <figcaption class="caption_text font_sf_text">
+          Image caption centered this way and I’ll make this a bit longer to indicate the
+          amount of line-height.
+        </figcaption>
+      </div>
 
-    <div class="article_block center media_mobile">
+      <div class="article_block center media_mobile">
+        <hr
+          style="
+            border: none;
+            background-color: black;
+            height: 2px;
+            color: black;
+            margin-bottom: 35px;
+          "
+        />
+        <span class="article_text font_normal font_ny_m">
+          <p>{{ currentPoster.description }}</p>
+        </span>
+      </div>
       <hr
         style="
-          border: none;
+          height: 12px;
           background-color: black;
-          height: 2px;
-          color: black;
-          margin-bottom: 35px;
+          margin-top: 77px;
+          min-width: 330px;
+          max-width: 640px;
         "
       />
-      <span class="article_text font_normal font_ny_m">
-        <p>{{ currentPoster.description }}</p>
-      </span>
-    </div>
-    <hr
-      style="
-        height: 12px;
-        background-color: black;
-        margin-top: 77px;
-        min-width: 330px;
-        max-width: 640px;
-      "
-    />
 
-    <div class="btn_group">
-      <div>
-        <button
-          type="button"
-          class="btn_action"
-          style="margin-top: 25px"
-          v-on:click="visible = true"
-        >
-          Update
-        </button>
+      <div class="btn_group">
+        <div>
+          <button
+            type="button"
+            class="btn_action"
+            style="margin-top: 25px"
+            v-on:click="visible = true"
+          >
+            Update
+          </button>
+        </div>
+        <div>
+          <button
+            type="button"
+            class="btn_action"
+            style="margin-top: 25px"
+            v-on:click="deletePoster"
+          >
+            Delete
+          </button>
+        </div>
       </div>
-      <div>
-        <button
-          type="button"
-          class="btn_action"
-          style="margin-top: 25px"
-          v-on:click="deletePoster"
-        >
-          Delete
-        </button>
-      </div>
+
+      <Alert v-if="getErrorMessage" />
+      <InputForm v-bind:screenSize="screenSize" />
     </div>
-    <InputForm v-bind:screenSize="screenSize" />
   </div>
 </template>
 
@@ -73,11 +80,16 @@ import Navigation from "@/components/Navigation";
 import Footer from "@/components/Footer";
 import InputForm from "@/components/InputForm";
 import Poster from "./Poster";
-import { mapGetters } from "vuex";
+import { mapGetters, mapMutations } from "vuex";
+import Alert from "../components/Alert";
+import Loader from "@/components/Loader";
+
 export default {
   components: {
     InputForm,
     Poster,
+    Alert,
+    Loader,
   },
   data() {
     return {
@@ -92,26 +104,34 @@ export default {
       if (this.screenSize > 645) this.sideMenu = false;
     },
     async deletePoster() {
-      await this.$store.dispatch("deletePoster", {
-        id: this.id,
-        currentPoster: this.currentPoster,
-      });
+      if (this.getUsername && this.currentPoster.owner !== this.getUsername) {
+        this.setErrorAlert("You don't have permissions to delete post");
+      } else {
+        this.clearPosters();
+        await this.$store.dispatch("deletePoster", {
+          id: this.id,
+          currentPoster: this.currentPoster,
+        });
 
-      setTimeout(() => {
-        this.$router.push("/");
-      }, 1000);
+        setTimeout(() => {
+          this.$router.push("/");
+        }, 1000);
+      }
     },
     closeModal() {
       this.visible = false;
     },
+    ...mapMutations(["setLastKey", "clearPosters", "setErrorAlert", "setLoading"]),
   },
   async mounted() {
     window.addEventListener("resize", this.onResize);
     this.onResize();
     this.id = this.$route.params.id;
+    this.setLoading(true);
     await this.$store.dispatch("getPosterById", this.id);
+    this.setLoading(false);
   },
-  computed: mapGetters(["currentPoster"]),
+  computed: mapGetters(["currentPoster", "getUsername", "getErrorMessage", "getLoading"]),
 };
 </script>
